refactor(platform-wallet-ffi): split platform_address_wallet.rs into matching file structure
Mirror the source structure of platform_wallet::wallet::platform_addresses:
- platform_addresses/mod.rs — shared runtime, input selection parsing, re-exports
- platform_addresses/wallet.rs — handle management, queries, free functions
- platform_addresses/sync.rs — sync_balances, sync_balances_on_account
- platform_addresses/transfer.rs — transfer
- platform_addresses/withdrawal.rs — withdraw
- platform_addresses/fund_from_asset_lock.rs — fund_from_asset_lock

Extracts parse_input_selection helper to mod.rs to avoid duplication
between transfer and withdrawal.
