+++ packages/rs-platform-wallet-ffi/src/lib.rs
@@ -12,7 +12,7 @@ pub mod handle;
 pub mod identity_manager;
 pub mod managed_identity;
 pub mod platform_address_types;
-pub mod platform_address_wallet;
+pub mod platform_addresses;
 pub mod platform_wallet_info;
 pub mod types;
 pub mod utils;
@@ -26,7 +26,7 @@ pub use handle::*;
 pub use identity_manager::*;
 pub use managed_identity::*;
 pub use platform_address_types::*;
-pub use platform_address_wallet::*;
+pub use platform_addresses::*;
 pub use platform_wallet_info::*;
 pub use types::*;
 pub use utils::*;
